Factor earthscope-specific filter workarounds out of mth5_helpers
- update import of repair_missing_filters
- This will make merging with fourier_coefficients branch easier
